[FIX] udes_sale_stock: Set sale_line is_cancelled on picking.action_cancel
Using api onchange is inefficient, it will calcualte for every state
change. Set this only when cancelling the final picking.
